[FIX] 파일 내용이 있을 때만 파일 삭제하기
안그러면 NPE 남

+++ src/main/java/com/sushiblog/backend/service/post/BlogServiceImpl.java
@@ -102,8 +102,10 @@ public class BlogServiceImpl implements BlogService {
 
         if(blog.getCategory().getUser() == user) {
             blogRepository.delete(blog);
-            File file = new File(blog.getFilePath());
-            file.delete();
+            if(blog.getFilePath() != null) {
+                File file = new File(blog.getFilePath());
+                file.delete();
+            }
         }
     }
